添加了所有模块的转让求购功能
1 添加了对3个数据库的Codefirst的测试代码（功能基本没太多问题）
2 修复了Logo的显示,首页采用大Logo
3 完善了所有模块的Mapping代码，Mysql数据库暂时还没有生成，后续会把脚本完善
4 优化MVC视图引擎的查找速度，代码于Framework中CsharpRazorViewEngiee
5 完善了所有的模块的发布代码

+++ Fx.InformationPlatform.Site/Fx.InformationPlatform.Site/ViewModel/TransferViewGoods.cs
@@ -7,13 +7,12 @@ using Fx.Entity;
 namespace Fx.InformationPlatform.Site.ViewModel
 {
     public class TransferViewGoods
-    {
-       
+    {       
         public string Title { get; set; }
 
         public int CatagroyId { get; set; }
 
-        public int Price { get; set; }
+        public decimal Price { get; set; }
 
         public bool IsChangeGoods { get; set; }
 
@@ -49,6 +48,5 @@ namespace Fx.InformationPlatform.Site.ViewModel
             OtherFiles = new List<TransferPicture>();
             BadFiles = new List<TransferPicture>();
         }
-
     }
 }
